refactor(eval): split transformations.rs into module directory
Modularize the 3379-line transformations.rs into focused submodules
for maintainability. No behavior changes; identical test results
(366 tests pass).

Also removes dead code in expand_placeholder_string: the
has_unresolved variable was only set when wildcard=false, making
the `has_unresolved && wildcard` branch unreachable.

New structure:
  transformations/
    mod.rs       -- Transformation enum and apply() dispatch
    helpers.rs   -- all transformation helper functions
    tests.rs     -- all unit tests
